Create buffer strategy only once
Changed the way the buffer strategy was created and checked every time. Removed the if statement in the render method and instead put it in the start method. The buffer strategy is only created once so there is no need to check if it is null and recreate it, which used to decrease performance.