Start writing Python API
Started writing a Python API which allows the developer to use Python and JSON to use the engine. I will also add options to read JSON and use that information to load a scene. The engine will be a standalone jar, along with the option to download/install the Python API package.